Изменил логику токенов для нескольких устройств

+++ app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -4,10 +4,12 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Auth\LoginRequest;
+use App\Models\ApiToken;
 use Illuminate\Http\Client\ConnectionException;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Http;
 use Illuminate\View\View;
 
 class AuthenticatedSessionController extends Controller
@@ -35,10 +37,13 @@ class AuthenticatedSessionController extends Controller
 
     /**
      * Destroy an authenticated session.
+     * @throws ConnectionException
      */
     public function destroy(Request $request): RedirectResponse
     {
-        $request->user()->delete();
+        Http::api()->post('/logout');
+        
+        ApiToken::where('user_id', $request->user()->id)->where('app_name', config('app.name'))->delete();
 
         Auth::guard('web')->logout();
